requirements.txt
asgiref==3.11.1
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.11.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
typing_extensions==4.15.0
